Use test-net1 by default for local VMs, provide helper text for picking addresses
RFC5737 defines several networks for "Documentation or source code" which was described as for test purposes. In this pull request, I suggest encouraging people to move to using test-net1 (192.0.2.0/24) for all locally hosted virtual machines (presumably for the purposes of testing source code?).

Just to be clear, I'm not at all precious about this, so I won't mind if it's rejected :) but I do think it's better to limit the scope of potential conflicts in your IP addressing by using ranges not recommended for public or private use.