Updating to Ruby 1.9.3 doesn't succeed
I noticed this yesterday after the Puppet 3.5 release when my VMs started failing. Since the -y switch isn't passed to yum and the output is passed to /dev/null, the Ruby 1.9.3 installation doesn't occur. Another thing I noticed is that, if I changed the file as above, ```gem``` and ```ruby``` are both not found immediately after the install. It was kind of weird because, when going to the next step (previously "librarian-puppet.sh", now "r10k.sh"), that ```gem``` was now available.

Why this is happening, I can't really say, but I can confirm that Ruby 1.9.3 was not installed. I checked Ruby's version and it was still on 1.8.

FYI, I'm running a CentOS VM, so it might be something specific to CentOS.