Use DigitalOcean token instead of api user/key
Related to #886 - I tried to run `vagrant up` on my machine using a DigitalOcean configuration, and it kept giving me the error:

```
Digital Ocean Provider:
* Token is required
```

Since I already updated the `config.yml` with my token, I was very confused. Changing these lines in the Vagrantfile to use the token instead of the client_id and api_key resolves this issue.