Shows all section information.
Instead of hiding a service's settings if it is not marked for
installation, this change shows all information by default,
whether service will be marked for installation or not.

Each service is already organized within tabs, additional hiding
of information serves no real purpose.

This also removes code that disables form elements when a
contradicting service is marked for installation.

For example, previously if HHVM were marked for installation, all
of the PHP service fields would be marked as disabled. This created
a problem as disabled fields are not POSTed on form submit and the
sections would not appear in the YAML config file.