Forces all extensions to submit information.
HTML forms will not submit an input[type=checkbox] if the value
is 0.

This breaks our process of throwing all extensions into a
zip because the puphpet engine doesn't see that any information
was sent for that extension.

When an extension install checkbox is not checked, add a hidden
input field immediately after with value "0". This overrides the
unchecked checkbox. When user checks the checkbox to mark
extension for installation, delete the hidden field. If they
uncheck the checkbox, readd the field.