Sets gui-config hostname as hidden #928
Resets hostname default value to ~ (null in yaml). The reasoning
for this is because if you are creating multiple VMs quickly and
forget to change the hostname, Things Will Break.

This is already the case with IP address, but I feel this field
is so important that a user know about and customize that it is
required to be visible and manageable by user via the gui.

Adding field to gui and hiding it allows user to set value in the
yaml, and keep its value across multiple drag/drops into the gui.