Ensure nginx is included in generated config.yaml
With 'install' set to '0', instead of missing altogether. Broke this in
325daa7. Need to make sure Apache is installed, and that's up next.

Here's the error from Puppet (via vagrant up) that tipped me off:

==> default: Error: $nginx_values is false, not a hash or array at /tmp/vagrant-puppet/manifests-893fcda21077f5a583c2dbbba8515c0a/nodes/Nginx.pp:167 on node localhost
==> default: Error: $nginx_values is false, not a hash or array at /tmp/vagrant-puppet/manifests-893fcda21077f5a583c2dbbba8515c0a/nodes/Nginx.pp:167 on node localhost
The SSH command responded with a non-zero exit status. Vagrant
assumes that this means the command failed. The output for this command
should be in the log above. Please read the output to determine what
went wrong.