Improve MariaDB tmpdir permission settings
As mentioned in #1805, MariaDB install will overwrite the permissions for the /tmp directory, making it so only the mysql user can use it. This is obviously an unexpected behavior and should be fixed. I'm unsure whether this should just check if the tmpdir is not /tmp or remove this check completely.